% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launchApp}
\usage{
launchApp()
}
\value{
Shiny web application
}
\description{
Launches Shiny web application using the server.R and ui.R files located in
the /application subdirectory.
}
