% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBox.R
\name{makeBox}
\alias{makeBox}
\title{makeBox}
\usage{
makeBox(shp, uid)
}
\arguments{
\item{shp}{The SVI shapefile from the CDC website}

\item{uid}{The click event polygon id}
}
\value{
A styled group of boxplots using the SVI shapefile \code{shp}
and click event polygon id \code{uid}
}
\description{
Creates and styles plotly box plots using leaflet user click events. These
boxplots visualize the spread of data for each of the thematic domains
and adds a point indicating where a selected census tract falls within
the nationwide dataset. This point updates dynamically based on click
events.
}
\examples{
makeBox(shp = uploadShpFile(), uid = 62878)
makeBox(shp = uploadShpFile(), uid = event$id)

}
