% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDash.R
\name{makeDash}
\alias{makeDash}
\title{makeDash}
\usage{
makeDash(shp, uid)
}
\arguments{
\item{shp}{The SVI shapefile from the CDC website}

\item{uid}{The click event polygon id}
}
\value{
A styled dashboard using the SVI shapefile \code{shp}
and click event polygon id \code{uid}
}
\description{
Creates and styles plotly dashboard using leaflet user click events. This
dashboard visualizes each of the indicator variables that comprise a thematic
domain, and updates dynamically based on click events.
}
\examples{
makeDash(shp = uploadShpFile(), uid = 62878)
makeDash(shp = uploadShpFile(), uid = event$id)

}
