% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStyle.R
\name{mapStyle}
\alias{mapStyle}
\title{mapStyle}
\usage{
mapStyle(map, data, rpl, flag)
}
\arguments{
\item{map}{A leaflet map object}

\item{data}{The SVI shapefile from the CDC website}

\item{rpl}{The thematic domain field from the SVI shapefile}

\item{flag}{The thematic flag field from the SVI shapefile}
}
\value{
A styled leaflet map based on \code{rpl} and \code{flag} values
}
\description{
Creates and styles leaflet maps using the same style definition for
uniform maps. More information on thematic domains and thematic flags can
be found at the CDC website.
}
